# lixos/bluestore, PoC, DNM: release disk extents in lazy and bulky manner