--- a/src/os/bluestore/BitMapAllocator.h
+++ b/src/os/bluestore/BitMapAllocator.h
@@ -36,8 +36,6 @@ public:
     int64_t hint, mempool::bluestore_alloc::vector<AllocExtent> *extents) override;
 
   void release(
-    uint64_t offset, uint64_t length) override;
-  void release(
     const interval_set<uint64_t>& release_set) override;
 
   uint64_t get_free() override;