--- a/src/os/bluestore/StupidAllocator.h
+++ b/src/os/bluestore/StupidAllocator.h
@@ -47,8 +47,6 @@ public:
     uint64_t *offset, uint32_t *length);
 
   void release(
-    uint64_t offset, uint64_t length) override;
-  void release(
     const interval_set<uint64_t>& release_set) override;
 
   uint64_t get_free() override;