Rough performance tests
----------------------------------

### After the change
```
Jobs: 20 (f=1280): [w(20)][14.3%][r=0KiB/s,w=729MiB/s][r=0,w=186k IOPS][eta 00m
Jobs: 20 (f=1280): [w(20)][19.0%][r=0KiB/s,w=726MiB/s][r=0,w=186k IOPS][eta 00m
Jobs: 20 (f=1280): [w(20)][23.8%][r=0KiB/s,w=720MiB/s][r=0,w=184k IOPS][eta 00m
Jobs: 20 (f=1280): [w(20)][28.6%][r=0KiB/s,w=724MiB/s][r=0,w=185k IOPS][eta 00m
Jobs: 20 (f=1280): [w(20)][33.3%][r=0KiB/s,w=702MiB/s][r=0,w=180k IOPS][eta 00m
Jobs: 20 (f=1280): [w(20)][38.1%][r=0KiB/s,w=722MiB/s][r=0,w=185k IOPS][eta 00m
Jobs: 20 (f=1280): [w(20)][42.9%][r=0KiB/s,w=730MiB/s][r=0,w=187k IOPS][eta 00m
Jobs: 20 (f=1280): [w(20)][47.6%][r=0KiB/s,w=733MiB/s][r=0,w=188k IOPS][eta 00m
Jobs: 20 (f=1280): [w(20)][52.4%][r=0KiB/s,w=719MiB/s][r=0,w=184k IOPS][eta 00m
Jobs: 20 (f=1280): [w(20)][57.1%][r=0KiB/s,w=725MiB/s][r=0,w=186k IOPS][eta 00m
Jobs: 20 (f=1280): [w(20)][61.9%][r=0KiB/s,w=708MiB/s][r=0,w=181k IOPS][eta 00m
Jobs: 20 (f=1280): [w(20)][66.7%][r=0KiB/s,w=735MiB/s][r=0,w=188k IOPS][eta 00m
Jobs: 20 (f=1280): [w(20)][71.4%][r=0KiB/s,w=730MiB/s][r=0,w=187k IOPS][eta 00m
Jobs: 20 (f=1280): [w(20)][76.2%][r=0KiB/s,w=722MiB/s][r=0,w=185k IOPS][eta 00m
Jobs: 20 (f=1280): [w(20)][81.0%][r=0KiB/s,w=704MiB/s][r=0,w=180k IOPS][eta 00m
Jobs: 20 (f=1280): [w(20)][85.7%][r=0KiB/s,w=695MiB/s][r=0,w=178k IOPS][eta 00m
Jobs: 20 (f=1280): [w(20)][90.5%][r=0KiB/s,w=709MiB/s][r=0,w=181k IOPS][eta 00m
Jobs: 20 (f=1280): [w(20)][95.2%][r=0KiB/s,w=726MiB/s][r=0,w=186k IOPS][eta 00m
Jobs: 20 (f=1280): [w(20)][100.0%][r=0KiB/s,w=723MiB/s][r=0,w=185k IOPS][eta 00

```

### Before the change
Tested on relatively fresh `master` (dd4cc66673df048d0840d31fb0bacc514ad481df, `Fri Sep 1 03:36:42 2017 +0000`).
```
Jobs: 20 (f=1280): [w(20)][14.3%][r=0KiB/s,w=497MiB/s][r=0,w=127k IOPS][eta 00m
Jobs: 20 (f=1280): [w(20)][19.0%][r=0KiB/s,w=481MiB/s][r=0,w=123k IOPS][eta 00m
Jobs: 20 (f=1280): [w(20)][23.8%][r=0KiB/s,w=504MiB/s][r=0,w=129k IOPS][eta 00m
Jobs: 20 (f=1280): [w(20)][28.6%][r=0KiB/s,w=466MiB/s][r=0,w=119k IOPS][eta 00m
Jobs: 20 (f=1280): [w(20)][33.3%][r=0KiB/s,w=501MiB/s][r=0,w=128k IOPS][eta 00m
Jobs: 20 (f=1280): [w(20)][38.1%][r=0KiB/s,w=467MiB/s][r=0,w=120k IOPS][eta 00m
Jobs: 20 (f=1280): [w(20)][42.9%][r=0KiB/s,w=490MiB/s][r=0,w=125k IOPS][eta 00m
Jobs: 20 (f=1280): [w(20)][47.6%][r=0KiB/s,w=497MiB/s][r=0,w=127k IOPS][eta 00m
Jobs: 20 (f=1280): [w(20)][52.4%][r=0KiB/s,w=492MiB/s][r=0,w=126k IOPS][eta 00m
Jobs: 20 (f=1280): [w(20)][57.1%][r=0KiB/s,w=479MiB/s][r=0,w=123k IOPS][eta 00m
Jobs: 20 (f=1280): [w(20)][61.9%][r=0KiB/s,w=473MiB/s][r=0,w=121k IOPS][eta 00m
Jobs: 20 (f=1280): [w(20)][66.7%][r=0KiB/s,w=484MiB/s][r=0,w=124k IOPS][eta 00m
Jobs: 20 (f=1280): [w(20)][71.4%][r=0KiB/s,w=487MiB/s][r=0,w=125k IOPS][eta 00m
Jobs: 20 (f=1280): [w(20)][76.2%][r=0KiB/s,w=485MiB/s][r=0,w=124k IOPS][eta 00m
Jobs: 20 (f=1280): [w(20)][81.0%][r=0KiB/s,w=488MiB/s][r=0,w=125k IOPS][eta 00m
Jobs: 20 (f=1280): [w(20)][85.7%][r=0KiB/s,w=491MiB/s][r=0,w=126k IOPS][eta 00m
Jobs: 20 (f=1280): [w(20)][90.5%][r=0KiB/s,w=480MiB/s][r=0,w=123k IOPS][eta 00m
Jobs: 20 (f=1280): [w(20)][95.2%][r=0KiB/s,w=487MiB/s][r=0,w=125k IOPS][eta 00m
Jobs: 20 (f=1280): [w(20)][100.0%][r=0KiB/s,w=476MiB/s][r=0,w=122k IOPS][eta 00
```

### The configuration
```
$ cat ceph-bluestore.fio 
# Runs a 64k random write test against the ceph BlueStore.
[global]
ioengine=libfio_ceph_objectstore.so # must be found in your LD_LIBRARY_PATH

conf=ceph-bluestore.conf # must point to a valid ceph configuration file
directory=/work/bluestore_testing # directory for osd_data

rw=randwrite
iodepth=16

time_based=1
runtime=20s
numjobs=20

[bluestore]
nr_files=64
size=256k
bs=4k
```

```
$ cat ceph-bluestore.conf 
# example configuration file for ceph-bluestore.fio

[global]
	debug bluestore = 0/0
	debug bluefs = 0/0
	debug bdev = 0/0
	debug rocksdb = 0/0
	# spread objects over 8 collections
	osd pool default pg num = 8
	# increasing shards can help when scaling number of collections
	osd op num shards = 5
	bluestore block create = false
	bluestore block path = /dev/nvme0n1p4
	bluestore_debug_omit_block_device_write = true
	bluestore_debug_omit_kv_commit = true
	bluestore_allocator = stupid # bitmap
	bluestore_min_alloc_size = 4096 #16384


[osd]
	osd objectstore = bluestore

	# use directory= option from fio job file
	osd data = ${fio_dir}

	# log inside fio_dir
	log file = ${fio_dir}/log
```