@liewegas 2017-09-07
The first two patches are a no-brainer; the third one makes me very nervous.

How much of the benefit can you get from just the first two?