@rzarzynski 2017-09-08
> How much of the benefit can you get from just the first two?

I'm afraid the lion's share comes from lazy releases. Below are results with the last commit reverted:

```
Jobs: 20 (f=1280): [w(20)][14.3%][r=0KiB/s,w=517MiB/s][r=0,w=132k IOPS][eta 00m
Jobs: 20 (f=1280): [w(20)][19.0%][r=0KiB/s,w=515MiB/s][r=0,w=132k IOPS][eta 00m
Jobs: 20 (f=1280): [w(20)][23.8%][r=0KiB/s,w=512MiB/s][r=0,w=131k IOPS][eta 00m
Jobs: 20 (f=1280): [w(20)][28.6%][r=0KiB/s,w=505MiB/s][r=0,w=129k IOPS][eta 00m
Jobs: 20 (f=1280): [w(20)][33.3%][r=0KiB/s,w=505MiB/s][r=0,w=129k IOPS][eta 00m
Jobs: 20 (f=1280): [w(20)][38.1%][r=0KiB/s,w=501MiB/s][r=0,w=128k IOPS][eta 00m
Jobs: 20 (f=1280): [w(20)][42.9%][r=0KiB/s,w=497MiB/s][r=0,w=127k IOPS][eta 00m
Jobs: 20 (f=1280): [w(20)][47.6%][r=0KiB/s,w=502MiB/s][r=0,w=129k IOPS][eta 00m
Jobs: 20 (f=1280): [w(20)][52.4%][r=0KiB/s,w=499MiB/s][r=0,w=128k IOPS][eta 00m
Jobs: 20 (f=1280): [w(20)][57.1%][r=0KiB/s,w=494MiB/s][r=0,w=126k IOPS][eta 00m
Jobs: 20 (f=1280): [w(20)][61.9%][r=0KiB/s,w=491MiB/s][r=0,w=126k IOPS][eta 00m
Jobs: 20 (f=1280): [w(20)][66.7%][r=0KiB/s,w=494MiB/s][r=0,w=127k IOPS][eta 00m
Jobs: 20 (f=1280): [w(20)][71.4%][r=0KiB/s,w=496MiB/s][r=0,w=127k IOPS][eta 00m
Jobs: 20 (f=1280): [w(20)][76.2%][r=0KiB/s,w=496MiB/s][r=0,w=127k IOPS][eta 00m
Jobs: 20 (f=1280): [w(20)][81.0%][r=0KiB/s,w=494MiB/s][r=0,w=126k IOPS][eta 00m
Jobs: 20 (f=1280): [w(20)][85.7%][r=0KiB/s,w=495MiB/s][r=0,w=127k IOPS][eta 00m
Jobs: 20 (f=1280): [w(20)][90.5%][r=0KiB/s,w=495MiB/s][r=0,w=127k IOPS][eta 00m
Jobs: 20 (f=1280): [w(20)][95.2%][r=0KiB/s,w=490MiB/s][r=0,w=125k IOPS][eta 00m
Jobs: 20 (f=1280): [w(20)][100.0%][r=0KiB/s,w=489MiB/s][r=0,w=125k IOPS][eta 00
```