@liewegas 2017-09-08
can you do a test with fio-librbd to see how much it helps there?  I'm guessing it won't be 2x!  :)