@liewegas 2018-10-29
@rzarzynski can we do a new pr with only the 2 original patches?  the other changes don't look safe.